requirements.txt
adal==1.2.7
azure-common==1.1.27
azure-core==1.20.1
azure-functions==1.8.0
azure-identity==1.7.1
azure-keyvault-secrets==4.3.0
azure-storage-blob==12.9.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cryptography==36.0.0
idna==3.3
isodate==0.6.0
lxml==4.6.4
msal==1.16.0
msal-extensions==0.3.0
msrest==0.6.21
msrestazure==0.6.4
multitasking==0.0.10
numpy==1.21.4
oauthlib==3.1.1
pandas==1.3.4
pandas-datareader==0.10.0
portalocker==1.7.1
pycparser==2.21
PyJWT==2.3.0
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
urllib3==1.26.7
yfinance==0.1.67
azure.cosmosdb.table

